+++ clfft2.cpp
@@ -109,7 +109,7 @@ int main(int argc, char* argv[]) {
   clfft2 fft(nx,ny,queue,ctx);
   fft.create_clbuf();
 
-  float *X = fft.create_rambuf();
+  double *X = fft.create_rambuf();
 
   std::cout << "\nInput:" << std::endl;
   init(X,nx,ny);
